Se agregó middleware para manejo global de errores y se corrigieron varios bugs de los endpoints en caso de datos incorrectos

+++ src/models/product.model.js
@@ -1,13 +1,13 @@
-import { 
-    collection, 
-    getDocs, 
-    getDoc, 
-    doc, 
-    query, 
-    where, 
-    addDoc, 
-    updateDoc, 
-    deleteDoc 
+import {
+    collection,
+    getDocs,
+    getDoc,
+    doc,
+    query,
+    where,
+    addDoc,
+    updateDoc,
+    deleteDoc
 } from 'firebase/firestore';
 import { db } from '../config/firebase.js';
 
@@ -31,25 +31,49 @@ const findByCategory = async (category) => {
 };
 
 const create = async (data) => {
-    const normalizedData = { 
-        ...data, 
-        category: data.category.toLowerCase() 
+    const normalizedData = {
+        ...data,
+        category: data.category.toLowerCase()
     };
     
-    const ref = await addDoc(collection(db, COLLECTION_NAME), normalizedData);
+    const q = query(
+        collection(db, COLLECTION_NAME),
+        where("name", "==", normalizedData.name),
+        where("category", "==", normalizedData.category)
+    );
 
+    const snapshot = await getDocs(q);
+
+    if (!snapshot.empty) {
+        
+        return null;
+    }
+
+    const ref = await addDoc(collection(db, COLLECTION_NAME), normalizedData);
     return { id: ref.id, ...normalizedData };
 };
 
 const updateById = async (id, data) => {
     const ref = doc(db, COLLECTION_NAME, id);
-    await updateDoc(ref, data);
     const snapshot = await getDoc(ref);
-    return { id: snapshot.id, ...snapshot.data() };
+
+    if (!snapshot.exists()) {
+        return null;
+    }
+
+    await updateDoc(ref, data);
+    const updatedSnapshot = await getDoc(ref);
+    return { id: updatedSnapshot.id, ...updatedSnapshot.data() };
 };
 
 const deleteById = async (id) => {
     const ref = doc(db, COLLECTION_NAME, id);
+    const snapshot = await getDoc(ref);
+
+    if (!snapshot.exists()) {
+        return false;
+    }
+
     await deleteDoc(ref);
     return true;
 };
